small api refactor
