+++ server/api/overzicht/producten/index.post.ts
@@ -1,15 +1,28 @@
 import { getMysqlConnection } from "~/server/mysql";
-import { ref } from "@vue/reactivity";
+
 export default defineEventHandler(async (event) => {
     const con = getMysqlConnection();
     const body = await readBody(event);
     try {
-        const search = body.searchTerm._value;
-        const [result] = await con
-            .promise()
-            .execute(`SELECT producten.*, categorie.naam AS categorie_naam FROM \`producten\` JOIN categorie ON categorie_id = categorie.id WHERE producten.naam LIKE '%${search.toLowerCase()}%' OR categorie.naam LIKE '%${search.toLowerCase()}%';`);
+        const [results, fields] = await con.promise().execute("INSERT INTO producten (ean, naam, categorie_id, aantal) VALUES (?, ?, ?, ?)", [body.ean, body.naam, body.categorie_id as Number, body.aantal]);
+        return {
+            status: 200,
+            body: results
+        }
+    } catch (error: any) {
+        let result = {
+            status: 500,
+            body: error
+        };
+        switch (error.errno) {
+            case 1062:
+                result.status = 409;
+                result.body = "EAN bestaat al";
+                break;
+            default:
+                break;
+        }
         return result;
-    } catch (error) {
-        return error;
+        
     }
 });
